fileHandling

+++ class 05/app.js
@@ -1,6 +1,7 @@
 import express from "express"
 import fs from "fs"
 import { json } from "stream/consumers"
+import { brotliDecompress } from "zlib"
 
 const app = express()
 const PORT = 5000
@@ -16,11 +17,36 @@ app.get("/", (req, res)=>{
 
 
 app.post("/createuser", (req,res)=>{
-    const tempArr =[]
-    tempArr.push(req.body)
-    console.log(tempArr,"tempARR")
-    fs.writeFileSync("users.txt", JSON.stringify(tempArr))
-    res.send("User Created!")
+    // const tempArr =[]
+    // tempArr.push(req.body)
+    // console.log(tempArr,"tempARR")
+    // fs.writeFileSync("users.txt", JSON.stringify(tempArr))
+    // res.send("User Created!")
+
+    //--------------NOW WORKING WITH CONDITIONS TO CHECK FILE EXIST OR NOT
+    //--------------IF EXSiST UPDATER ELSE CREAT----//
+
+    const fileExist = fs.existsSync("users.txt")
+
+    if(fileExist){
+        //APPEND
+        const getData = fs.readFileSync("users.txt", "utf-8")
+        const parseData = JSON.parse(getData)
+        parseData.push(req.body)
+        fs.writeFileSync("users.txt", JSON.stringify(parseData))
+        console.log("parse data" , parseData)
+        
+    } else{
+        //CREATE NEW FILE
+        const tempArr = []
+        tempArr.push(req.body)
+        fs.writeFileSync("users.txt", JSON.stringify(tempArr))
+    }
+    res.send("File created")
+
+
+
+
 })
